+++ source/renderer/app/components/staking/countdown/StakingCountdown.js
@@ -4,8 +4,8 @@ import { observer } from 'mobx-react';
 import { defineMessages, intlShape } from 'react-intl';
 import { Button } from 'react-polymorph/lib/components/Button';
 import { ButtonSkin } from 'react-polymorph/lib/skins/simple/ButtonSkin';
-import DelegationCountdownCounter from './DelegationCountdownCounter';
-import styles from './StakingDelegationCountdown.scss';
+import CountdownWidget from '../../widgets/CountdownWidget';
+import styles from './StakingCountdown.scss';
 
 const messages = defineMessages({
   heading: {
@@ -38,7 +38,7 @@ type Props = {
 };
 
 @observer
-export default class StakingDelegationCountdown extends Component<Props> {
+export default class StakingCountdown extends Component<Props> {
   static contextTypes = {
     intl: intlShape.isRequired,
   };
@@ -57,7 +57,7 @@ export default class StakingDelegationCountdown extends Component<Props> {
           <div className={styles.heading}>{heading}</div>
           <div className={styles.description}>{description}</div>
           <div className={styles.timeLeftDesc}>{timeLeftDesc}</div>
-          <DelegationCountdownCounter
+          <CountdownWidget
             redirectToStakingInfo={redirectToStakingInfo}
             currentLocale={currentLocale}
             startDateTime={startDateTime}