+++ source/renderer/app/stores/AppStore.js
@@ -5,13 +5,14 @@ import Store from './lib/Store';
 import LocalizableError from '../i18n/LocalizableError';
 import { buildRoute } from '../utils/routing';
 import { ROUTES } from '../routes-config';
-import { DIALOGS, SCREENS } from '../../../common/ipc/constants';
+import { DIALOGS, SCREENS, NOTIFICATIONS } from '../../../common/ipc/constants';
 import { openExternalUrlChannel } from '../ipc/open-external-url';
 import {
   toggleUiPartChannel,
   showUiPartChannel,
 } from '../ipc/control-ui-parts';
 import { getGPUStatusChannel } from '../ipc/get-gpu-status.ipc';
+import { generateFileNameWithTimestamp } from '../../../common/utils/files';
 
 import type { GpuStatus } from '../types/gpuStatus';
 
@@ -19,6 +20,7 @@ export default class AppStore extends Store {
   @observable error: ?LocalizableError = null;
   @observable isAboutDialogOpen = false;
   @observable isDaedalusDiagnosticsDialogOpen = false;
+  @observable isDownloadNotificationVisible = false;
   @observable gpuStatus: ?GpuStatus = null;
   @observable numberOfEpochsConsolidated: number = 0;
   @observable previousRoute: string = ROUTES.ROOT;
@@ -36,6 +38,12 @@ export default class AppStore extends Store {
     this.actions.app.getGpuStatus.listen(this._getGpuStatus);
     this.actions.app.toggleBlockConsolidationStatusScreen.listen(
       this._toggleBlockConsolidationStatusScreen
+    );
+
+    this.actions.app.downloadLogs.listen(this._downloadLogs);
+
+    this.actions.app.setNotificationVisibility.listen(
+      this._setDownloadNotification
     );
 
     toggleUiPartChannel.onReceive(this.toggleUiPart);
@@ -81,6 +89,9 @@ export default class AppStore extends Store {
     switch (uiPart) {
       case SCREENS.ADA_REDEMPTION:
         this._showAdaRedemptionScreen();
+        break;
+      case NOTIFICATIONS.DOWNLOAD_LOGS:
+        this._downloadLogs();
         break;
       default:
     }
@@ -156,5 +167,37 @@ export default class AppStore extends Store {
       ? this.previousRoute
       : ROUTES.BLOCK_CONSOLIDATION_STATUS;
     this._updateRouteLocation({ route });
+  };
+
+  @action _downloadLogs = () => {
+    if (this.isDownloadNotificationVisible) {
+      return;
+    }
+    const fileName = generateFileNameWithTimestamp();
+    global.dialog.showSaveDialog(
+      {
+        defaultPath: fileName,
+      },
+      destination => {
+        if (destination) {
+          this.actions.profile.downloadLogs.trigger({
+            fileName,
+            destination,
+            fresh: true,
+          });
+        } else {
+          this.actions.app.setNotificationVisibility.trigger(
+            !this.isDownloadNotificationVisible
+          );
+        }
+      }
+    );
+    this.isDownloadNotificationVisible = true;
+  };
+
+  @action _setDownloadNotification = (
+    isDownloadNotificationVisible: boolean
+  ) => {
+    this.isDownloadNotificationVisible = isDownloadNotificationVisible;
   };
 }