+++ source/renderer/app/themes/daedalus/dark-blue.js
@@ -365,6 +365,17 @@ const DARK_BLUE_THEME_CONFIG = {
     '--theme-loading-status-icons-tooltip-color': '#4b5a68',
     '--theme-loading-spinner-color': '#e9f4fe',
   },
+  manualUpdate: {
+    '--theme-manual-update-overlay-background-color': 'rgba(38, 51, 69, 0.96)',
+    '--theme-manual-update-overlay-button-background-color':
+      'rgba(0, 0, 0, 0.1)',
+    '--theme-manual-update-overlay-button-background-color-hover': '#fafbfc',
+    '--theme-manual-update-overlay-button-text-color-hover': '#263345',
+    '--theme-manual-update-overlay-button-border-color': '#fafbfc',
+    '--theme-manual-update-overlay-text-color': 'rgba(250, 251, 252, 0.7)',
+    '--theme-manual-update-overlay-text-highlight-color': '#fafbfc',
+    '--theme-manual-update-overlay-title-text-color': '#fafbfc',
+  },
   mnemonic: {
     '--theme-backup-mnemonic-background-color': 'rgba(233, 244, 254, 0.05)',
     '--theme-mnemonic-background-color': 'rgba(83, 99, 112, 0.3)',
@@ -421,6 +432,23 @@ const DARK_BLUE_THEME_CONFIG = {
   notification: {
     '--theme-notification-message-background-color': 'rgba(83, 99, 112, 0.88)',
     '--theme-notification-message-text-color': '#fafbfc',
+    '--theme-legacy-badge-background-color': '#ab1700',
+    '--theme-legacy-notification-background-color': '#ab2712',
+    '--theme-legacy-notification-learn-more-button-text-color': '#fafbfc',
+    '--theme-legacy-notification-learn-more-button-background-color':
+      'rgba(250, 251, 252, 0.2)',
+    '--theme-legacy-notification-learn-more-button-background-color-hover':
+      'rgba(250, 251, 252, 0.3)',
+    '--theme-legacy-notification-learn-more-button-background-color-active':
+      'rgba(250, 251, 252, 0.4)',
+    '--theme-legacy-notification-move-button-text-color': '#ab2712',
+    '--theme-legacy-notification-move-button-background-color': '#fafbfc',
+    '--theme-legacy-notification-move-button-background-color-hover':
+      'rgba(250, 251, 252, 0.9)',
+    '--theme-legacy-notification-move-button-background-color-active':
+      'rgba(250, 251, 252, 0.8)',
+    '--theme-legacy-notification-title-color': '#fafbfc',
+    '--theme-legacy-notification-description-color': '#fafbfc',
   },
   paperWallet: {
     '--theme-paper-wallet-create-certificate-dialog-explorer-link-color':