+++ source/renderer/app/themes/daedalus/light-blue.js
@@ -362,6 +362,16 @@ const LIGHT_BLUE_THEME_CONFIG = {
     '--theme-loading-status-icons-tooltip-color': '#062148',
     '--theme-loading-spinner-color': '#5e6066',
   },
+  manualUpdate: {
+    '--theme-manual-update-overlay-background-color': 'rgba(36, 62, 98, 0.96)',
+    '--theme-manual-update-overlay-button-background-color': '#243e62',
+    '--theme-manual-update-overlay-button-background-color-hover': '#fafbfc',
+    '--theme-manual-update-overlay-button-border-color': '#fafbfc',
+    '--theme-manual-update-overlay-button-text-color-hover': '#243e62',
+    '--theme-manual-update-overlay-text-color': 'rgba(250, 251, 252, 0.7)',
+    '--theme-manual-update-overlay-text-highlight-color': '#fafbfc',
+    '--theme-manual-update-overlay-title-text-color': '#fafbfc',
+  },
   mnemonic: {
     '--theme-mnemonic-background-color': '#e6e9ee',
     '--theme-mnemonic-background-color-hover': '#eaecf1',
@@ -419,6 +429,23 @@ const LIGHT_BLUE_THEME_CONFIG = {
   notification: {
     '--theme-notification-message-background-color': 'rgba(36, 61, 98, 0.88)',
     '--theme-notification-message-text-color': '#fafbfc',
+    '--theme-legacy-badge-background-color': '#ab1700',
+    '--theme-legacy-notification-background-color': '#ab2712',
+    '--theme-legacy-notification-learn-more-button-text-color': '#fafbfc',
+    '--theme-legacy-notification-learn-more-button-background-color':
+      'rgba(250, 251, 252, 0.2)',
+    '--theme-legacy-notification-learn-more-button-background-color-hover':
+      'rgba(250, 251, 252, 0.3)',
+    '--theme-legacy-notification-learn-more-button-background-color-active':
+      'rgba(250, 251, 252, 0.4)',
+    '--theme-legacy-notification-move-button-text-color': '#ab2712',
+    '--theme-legacy-notification-move-button-background-color': '#fafbfc',
+    '--theme-legacy-notification-move-button-background-color-hover':
+      'rgba(250, 251, 252, 0.9)',
+    '--theme-legacy-notification-move-button-background-color-active':
+      'rgba(250, 251, 252, 0.8)',
+    '--theme-legacy-notification-title-color': '#fafbfc',
+    '--theme-legacy-notification-description-color': '#fafbfc',
   },
   paperWallet: {
     '--theme-paper-wallet-create-certificate-dialog-explorer-link-color':