+++ storybook/stories/support/StoryProvider.js
@@ -20,6 +20,7 @@ export const WALLETS = [
     assurance: WalletAssuranceModeOptions.NORMAL,
     hasPassword: false,
     passwordUpdateDate: new Date(),
+    isLegacy: false,
   },
   {
     id: '1',
@@ -30,6 +31,16 @@ export const WALLETS = [
     passwordUpdateDate: moment()
       .subtract(1, 'month')
       .toDate(),
+    isLegacy: false,
+  },
+  {
+    id: '2',
+    name: 'Legacy',
+    amount: new BigNumber(55.555),
+    assurance: WalletAssuranceModeOptions.NORMAL,
+    hasPassword: false,
+    passwordUpdateDate: new Date(),
+    isLegacy: true,
   },
 ];